minor style fixes

+++ assets/php/delete_folder.php
@@ -15,7 +15,9 @@
             rmdir($dirPath);
             header ('Location: ../../index.php');
         } 
-        unlink('../../'.$_SESSION['filename'].'.php');
+        if($_SESSION['filename'] !== 'index') {
+            unlink('../../'.$_SESSION['filename'].'.php');
+        }
     }
 
     function deleteInnerFiles($path) {
